% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced.matrix.functions.R
\name{matrices.shift.apply.function}
\alias{matrices.shift.apply.function}
\title{Shift two matrices against each other and apply function to the overlap}
\usage{
matrices.shift.apply.function(mbase, mshift, func, keepbasedim = FALSE,
  radial = FALSE, na.value = -99999)
}
\arguments{
\item{mshift, mbase}{matrices of to be shifted against each other. \code{NA}s are allowed, see argument \code{na.value}.}

\item{func}{character. One of "rmse" or "multav". See details.}

\item{keepbasedim}{logical. Only shift \code{mshift} as much that the center stays inside \code{mbase}.
This results in an output matrix of equal shape like \code{mbase}. Handy for sliding means. Defaults to FALSE.}

\item{radial}{logical. Only shift \code{mshift} up to a circle-like shape against \code{mbase}. 
That is, The center of \code{mshift} won't exceed the biggest ellipsis fitting into the center of \code{mbase}. 
This results in an output matrix of equal shape like \code{mbase} with \code{NA}s in the edges. Handy for circular images.}

\item{na.value}{integer. The value used internally in FORTRAN for \code{NA}s. Make sure this value does not appear in neither \code{mbase} nor \code{mshift}.
Defaults to \code{-99999}.}
}
\value{
a matrix filled with the results of \code{func} and row- and colnames that indicate the shift of \code{mshift} against \code{mbase}.
}
\description{
Shift two matrices against each other and apply a function to the overlap
}
\details{
Matrix \code{mshift} is shifted against \code{mbase} and the overlapping matrices passed to \code{fun}.
 The return value of selected function \code{func} is put into the resulting matrix at the position of the shift.
 
 This function uses a FORTRAN implementation for efficiency.
 
 The \code{func} argument specifies which type of function is to be 
 applied to the overlap. It can be:
 \itemize{
 	\item \code{rmse}: The Root Mean Square Error RMSE is calculated on the overlap.
 				Handy for finding the shift of the maximum resemblance of \code{mbase} and \code{mshift}.
 	\item \code{multav}: The overlapping matrix parts are multiplicated elementwise and then averaged.
 	      Handy for sliding means.
 }
}

