% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advanced.matrix.functions.R
\name{matrix.shift.bounds}
\alias{matrix.shift.bounds}
\title{Remaining bounds after shifting two matrices against each other}
\usage{
matrix.shift.bounds(dimbase, dimshift, rowshift, colshift)
}
\arguments{
\item{dimbase}{Dimension of static matrix, as returned by \code{\link{dim}}.}

\item{dimshift}{Dimension of shifted matrix, as returned by \code{\link{dim}}.}

\item{rowshift}{integer. column shift. See details.}

\item{colshift}{integer. row shift. See details.}
}
\value{
vector of bounds in the following order:
 \itemize{
 \item{\code{mirosh}: lower rowbound of \code{mshift}}
 \item{\code{marosh}: upper rowbound of \code{mshift}}
 \item{\code{micosh}: lower columnbound of \code{mshift}}
 \item{\code{macosh}: upper columnbound of \code{mshift}}
 \item{\code{miroba}: lower rowbound of \code{mbase}}
 \item{\code{maroba}: upper rowbound of \code{mbase}}
 \item{\code{micoba}: lower columnbound of \code{mbase}}
 \item{\code{macoba}: upper columnbound of \code{mbase}}
 }
}
\description{
Determine the indices that limit the remainig 
 matrix after shifting two matrices against each other.
}
\details{
Imagine two matrices \code{mshift} and \code{mbase}. You shift matrix \code{mshift} against \code{mbase} by 
 a certain amount of rows and columns. Now you have a remaining 
 matrix that is possibly smaller than \code{mshift} or \code{mbase}. The bounds of this 
 remaining matrix are in both matrices.\cr\cr
 This function returns only the bounds of the matrices \code{mshift} and \code{mbase}
 that limit the remainig matrix.
 \cr\cr
 The \code{rowshift} and \code{colshift} parameters describe the shifts in units of rows and columns respectively. 
 A positive value indicates an \strong{downwards} (!) / rightwards shift of matrix \code{m1} against \code{m2}.
 A negative value indicates a \strong{upwards} (!) / leftwards shift. Values greater than 
 the respective dimensions \code{dim} lead to indefinite bounds because there 
 is no remaining matrix after shifting. In this case, all bounds are set to \code{NA}.
 \cr\cr
}
\examples{
# Two equal 3x3-matrices
(m1 <- m2 <- matrix(1:9,3,3))
#      [,1] [,2] [,3]
# [1,]    1    4    7
# [2,]    2    5    8
# [3,]    3    6    9
 
# Shift by 0 rows and 0 columns
(bounds = matrix.shift.bounds(dim(m1),dim(m2),0,0))
# [1] 1 3 1 3 1 3 1 3
# Remaining part of m1:
 m1[bounds[1]:bounds[2],bounds[3]:bounds[4]]
#      [,1] [,2] [,3]
# [1,]    1    4    7
# [2,]    2    5    8
# [3,]    3    6    9
# Everything! Because nothing was shifted!
}

