% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitive.matrix.functions.R
\name{rotate.left}
\alias{rotate.left}
\title{Rotate matrix 90 degrees counter-clockwise}
\usage{
rotate.left(x)
}
\arguments{
\item{x}{a 2d-matrix}
}
\value{
the given matrix rotated 90 degrees counter-clockwise
}
\description{
Rotate matrix 90 degrees counter-clockwise
}
\details{
\code{t(apply(x,2,rev))}
}
\examples{
a<-matrix(1:9,3,3)
a
#     [,1] [,2] [,3]
#[1,]    1    4    7
#[2,]    2    5    8
#[3,]    3    6    9
rotate.left(a)
#     [,1] [,2] [,3]
#[1,]    7    8    9
#[2,]    4    5    6
#[3,]    1    2    3
}

